prime or not algo added
