+++ algorithms/Java/Maths/check-prime-or-not.java
@@ -1,0 +1,19 @@
+import java.util.*;
+
+public class main {
+    public static void main(String[] args) {
+        Scanner scn = new Scanner(System.in);
+        int num = scn.nextInt();
+        boolean ans = isPrime(num);
+        System.out.println(ans!=true?"not-prime":"prime");
+    }
+
+    public static boolean isPrime(int num) {
+        for (int i = 2; i * i <= num; i++) {
+            if (num % 2 == 0) {
+                return false;
+            }
+        }
+        return true;
+    }
+}
